cleanup getRemoteInfo2
