+++ __tests__/test-getRemoteInfo2.js
@@ -37,17 +37,21 @@ describe('getRemoteInfo2', () => {
     if (info.protocolVersion === 1) {
       expect(info.refs).toBeDefined()
       expect(info.refs).toMatchInlineSnapshot(`
-        Object {
-          "HEAD": "97c024f73eaab2781bf3691597bc7c833cb0e22f",
-          "refs/heads/master": "97c024f73eaab2781bf3691597bc7c833cb0e22f",
-          "refs/heads/test": "5a8905a02e181fe1821068b8c0f48cb6633d5b81",
-        }
-      `)
-      expect(info.symrefs).toBeDefined()
-      expect(info.symrefs).toMatchInlineSnapshot(`
-        Object {
-          "HEAD": "refs/heads/master",
-        }
+        Array [
+          Object {
+            "oid": "97c024f73eaab2781bf3691597bc7c833cb0e22f",
+            "ref": "HEAD",
+            "target": "refs/heads/master",
+          },
+          Object {
+            "oid": "97c024f73eaab2781bf3691597bc7c833cb0e22f",
+            "ref": "refs/heads/master",
+          },
+          Object {
+            "oid": "5a8905a02e181fe1821068b8c0f48cb6633d5b81",
+            "ref": "refs/heads/test",
+          },
+        ]
       `)
     }
   })
