+++ src/api/getRemoteInfo2.js
@@ -3,32 +3,29 @@ import '../typedefs.js'
 
 import { GitRemoteHTTP } from '../managers/GitRemoteHTTP.js'
 import { assertParameter } from '../utils/assertParameter.js'
-import { fromEntries } from '../utils/fromEntries.js'
+import { formatInfoRefs } from '../utils/formatInfoRefs.js'
 
 /**
- * @typedef {Object} GetRemoteInfoResult1 - This object has the following schema:
- * @property {1} protocolVersion - Git Protocol Version 1
- * @property {string[]} capabilities - A list of capabilities
- * @property {Object<string, string>} refs - An object of remote refs and corresponding SHA-1 object ids
- * @property {Object<string, string>} symrefs - An object of remote symrefs and corresponding refs
- */
-
-/**
- * @typedef {Object} GetRemoteInfoResult2 - This object has the following schema:
- * @property {2} protocolVersion - Git Protocol version 2
- * @property {Object<string, string | null>} capabilities - An object of capabilities represented as keys and values
+ * @typedef {Object} GetRemoteInfo2Result - This object has the following schema:
+ * @property {1 | 2} protocolVersion - Git protocol version the server supports
+ * @property {Object<string, string | true>} capabilities - An object of capabilities represented as keys and values
+ * @property {ServerRef[]} [refs] - Server refs (they get returned by protocol version 1 whether you want them or not)
  */
 
 /**
  * List a remote server's capabilities.
  *
- * > The successor to `getRemoteInfo`, this command supports Git Wire Protocol Version 2.
- * > Therefore its return type is more complicated, as *either* a v1 or v2 result is returned.
- * > Also, I've "fixed" the v1 return result so its a flat list of refs.
- * > The nested object thing looked nice to the human eye, but is a pain to deal with programatically.
- *
  * This is a rare command that doesn't require an `fs`, `dir`, or even `gitdir` argument.
  * It just communicates to a remote git server, determining what protocol version, commands, and features it supports.
+ *
+ * > The successor to [`getRemoteInfo`](./getRemoteInfo.md), this command supports Git Wire Protocol Version 2.
+ * > Therefore its return type is more complicated as either:
+ * >
+ * > - v1 capabilities (and refs) or
+ * > - v2 capabilities (and no refs)
+ * >
+ * > are returned.
+ * > If you just care about refs, use [`listServerRefs`](./listServerRefs.md)
  *
  * @param {object} args
  * @param {HttpClient} args.http - an HTTP client
@@ -41,9 +38,9 @@ import { fromEntries } from '../utils/fromEntries.js'
  * @param {Object<string, string>} [args.headers] - Additional headers to include in HTTP requests, similar to git's `extraHeader` config
  * @param {1 | 2} [args.protocolVersion = 2] - Which version of the Git Protocol to use.
  *
- * @returns {Promise<GetRemoteInfoResult1 | GetRemoteInfoResult2>} Resolves successfully with an object listing the branches, tags, and capabilities of the remote.
- * @see GetRemoteInfoResult1
- * @see GetRemoteInfoResult2
+ * @returns {Promise<GetRemoteInfo2Result>} Resolves successfully with an object listing the capabilities of the remote.
+ * @see GetRemoteInfo2Result
+ * @see ServerRef
  *
  * @example
  * let info = await git.getRemoteInfo2({
@@ -82,7 +79,7 @@ export async function getRemoteInfo2({
     })
 
     if (remote.protocolVersion === 2) {
-      /** @type GetRemoteInfoResult2 */
+      /** @type GetRemoteInfo2Result */
       return {
         protocolVersion: remote.protocolVersion,
         capabilities: remote.capabilities2,
@@ -92,12 +89,21 @@ export async function getRemoteInfo2({
     // Note: remote.capabilities, remote.refs, and remote.symrefs are Set and Map objects,
     // but one of the objectives of the public API is to always return JSON-compatible objects
     // so we must JSONify them.
-    /** @type GetRemoteInfoResult1 */
+    /** @type Object<string, true> */
+    const capabilities = {}
+    for (const cap of remote.capabilities) {
+      const [key, value] = cap.split('=')
+      if (value) {
+        capabilities[key] = value
+      } else {
+        capabilities[key] = true
+      }
+    }
+    /** @type GetRemoteInfo2Result */
     return {
       protocolVersion: 1,
-      capabilities: [...remote.capabilities],
-      refs: fromEntries(remote.refs),
-      symrefs: fromEntries(remote.symrefs),
+      capabilities,
+      refs: formatInfoRefs(remote, undefined, true, true),
     }
   } catch (err) {
     err.caller = 'git.getRemoteInfo2'
