+++ src/api/listServerRefs.js
@@ -3,22 +3,44 @@ import '../typedefs.js'
 
 import { GitRemoteHTTP } from '../managers/GitRemoteHTTP.js'
 import { assertParameter } from '../utils/assertParameter.js'
+import { formatInfoRefs } from '../utils/formatInfoRefs.js'
 import { parseListRefsResponse } from '../wire/parseListRefsResponse.js'
 import { writeListRefsRequest } from '../wire/writeListRefsRequest.js'
-
-/**
- * @typedef {Object} ServerRef - This object has the following schema:
- * @property {string} ref - The name of the ref
- * @property {string} oid - The SHA-1 object id the ref points to
- * @property {string} [target] - The target ref pointed to by a symbolic ref
- * @property {string} [peeled] - If the oid is the SHA-1 object id of an annotated tag, this is the SHA-1 object id that the annotated tag points to
- */
 
 /**
  * Fetch a list of refs (branches, tags, etc) from a server.
  *
  * This is a rare command that doesn't require an `fs`, `dir`, or even `gitdir` argument.
  * It just requires an `http` argument.
+ *
+ * ### About `protocolVersion`
+ *
+ * There's a rather fun trade-off between Git Protocol Version 1 and Git Protocol Version 2.
+ * Version 2 actually requires 2 HTTP requests instead of 1, making it similar to fetch or push in that regard.
+ * However, version 2 supports server-side filtering by prefix, whereas that filtering is done client-side in version 1.
+ * Which protocol is most efficient therefore depends on the number of refs on the remote, the latency of the server, and speed of the network connection.
+ * For an small repos (or fast Internet connections), the requirement to make two trips to the server makes protocol 2 slower.
+ * But for large repos (or slow Internet connections), the decreased payload size of the second request makes up for the additional request.
+ *
+ * Hard numbers vary by situation, but here's some numbers from my machine:
+ *
+ * Using isomorphic-git in a browser, with a CORS proxy, listing only the branches (refs/heads) of https://github.com/isomorphic-git/isomorphic-git
+ * - Protocol Version 1 took ~300ms and transfered 84 KB.
+ * - Protocol Version 2 took ~500ms and transfered 4.1 KB.
+ *
+ * Using isomorphic-git in a browser, with a CORS proxy, listing only the branches (refs/heads) of https://gitlab.com/gitlab-org/gitlab
+ * - Protocol Version 1 took ~4900ms and transfered 9.41 MB.
+ * - Protocol Version 2 took ~1280ms and transfered 433 KB.
+ *
+ * Finally, there is a fun quirk regarding the `symrefs` parameter.
+ * Protocol Version 1 will generally only return the `HEAD` symref and not others.
+ * Historically, this meant that servers don't use symbolic refs except for `HEAD`, which is used to point at the "default branch".
+ * However Protocol Version 2 can return *all* the symbolic refs on the server.
+ * So if you are running your own git server, you could take advantage of that I guess.
+ *
+ * #### TL;DR
+ * If you are _not_ taking advantage of `prefix` I would recommend `protocolVersion: 1`.
+ * Otherwise, I recommend to use the default which is `protocolVersion: 2`.
  *
  * @param {object} args
  * @param {HttpClient} args.http - an HTTP client
@@ -38,10 +60,44 @@ import { writeListRefsRequest } from '../wire/writeListRefsRequest.js'
  * @see ServerRef
  *
  * @example
+ * // List all the branches on a repo
  * let refs = await git.listServerRefs({
  *   http,
  *   corsProxy: "https://cors.isomorphic-git.org",
- *   url: "https://github.com/isomorphic-git/isomorphic-git.git"
+ *   url: "https://github.com/isomorphic-git/isomorphic-git.git",
+ *   prefix: "refs/heads/",
+ * });
+ * console.log(refs);
+ *
+ * @example
+ * // Get the default branch on a repo
+ * let refs = await git.listServerRefs({
+ *   http,
+ *   corsProxy: "https://cors.isomorphic-git.org",
+ *   url: "https://github.com/isomorphic-git/isomorphic-git.git",
+ *   prefix: "HEAD",
+ *   symrefs: true,
+ * });
+ * console.log(refs);
+ *
+ * @example
+ * // List all the tags on a repo
+ * let refs = await git.listServerRefs({
+ *   http,
+ *   corsProxy: "https://cors.isomorphic-git.org",
+ *   url: "https://github.com/isomorphic-git/isomorphic-git.git",
+ *   prefix: "refs/tags/",
+ *   peelTags: true,
+ * });
+ * console.log(refs);
+ *
+ * @example
+ * // List all the pull requests on a repo
+ * let refs = await git.listServerRefs({
+ *   http,
+ *   corsProxy: "https://cors.isomorphic-git.org",
+ *   url: "https://github.com/isomorphic-git/isomorphic-git.git",
+ *   prefix: "refs/pull/",
  * });
  * console.log(refs);
  *
@@ -77,33 +133,7 @@ export async function listServerRefs({
     })
 
     if (remote.protocolVersion === 1) {
-      const refs = []
-      for (const [key, value] of remote.refs) {
-        if (prefix && !key.startsWith(prefix)) continue
-
-        if (key.endsWith('^{}')) {
-          if (peelTags) {
-            const _key = key.replace('^{}', '')
-            // Peeled tags are almost always listed immediately after the original tag
-            const last = refs[refs.length - 1]
-            const r = last.ref === _key ? last : refs.find(x => x.ref === _key)
-            if (r === undefined) {
-              throw new Error('I did not expect this to happen')
-            }
-            r.peeled = value
-          }
-          continue
-        }
-        /** @type ServerRef */
-        const ref = { ref: key, oid: value }
-        if (symrefs) {
-          if (remote.symrefs.has(key)) {
-            ref.target = remote.symrefs.get(key)
-          }
-        }
-        refs.push(ref)
-      }
-      return refs
+      return formatInfoRefs(remote, prefix, symrefs, peelTags)
     }
 
     // Protocol Version 2
