+++ website/sidebars.json
@@ -62,7 +62,8 @@
       "getRemoteInfo2",
       "addRemote",
       "deleteRemote",
-      "listRemotes"
+      "listRemotes",
+      "listServerRefs"
     ],
     "Objects": [
       "readBlob",
