+++ website/versioned_docs/version-1.x/listServerRefs.md
@@ -37,13 +37,79 @@ type ServerRef = {
 This is a rare command that doesn't require an `fs`, `dir`, or even `gitdir` argument.
 It just requires an `http` argument.
 
+### About `protocolVersion`
+
+There's a rather fun trade-off between Git Protocol Version 1 and Git Protocol Version 2.
+Version 2 actually requires 2 HTTP requests instead of 1, making it similar to fetch or push in that regard.
+However, version 2 supports server-side filtering by prefix, whereas that filtering is done client-side in version 1.
+Which protocol is most efficient therefore depends on the number of refs on the remote, the latency of the server, and speed of the network connection.
+For an small repos (or fast Internet connections), the requirement to make two trips to the server makes protocol 2 slower.
+But for large repos (or slow Internet connections), the decreased payload size of the second request makes up for the additional request.
+
+Hard numbers vary by situation, but here's some numbers from my machine:
+
+Using isomorphic-git in a browser, with a CORS proxy, listing only the branches (refs/heads) of https://github.com/isomorphic-git/isomorphic-git
+- Protocol Version 1 took ~300ms and transfered 84 KB.
+- Protocol Version 2 took ~500ms and transfered 4.1 KB.
+
+Using isomorphic-git in a browser, with a CORS proxy, listing only the branches (refs/heads) of https://gitlab.com/gitlab-org/gitlab
+- Protocol Version 1 took ~4900ms and transfered 9.41 MB.
+- Protocol Version 2 took ~1280ms and transfered 433 KB.
+
+Finally, there is a fun quirk regarding the `symrefs` parameter.
+Protocol Version 1 will generally only return the `HEAD` symref and not others.
+Historically, this meant that servers don't use symbolic refs except for `HEAD`, which is used to point at the "default branch".
+However Protocol Version 2 can return *all* the symbolic refs on the server.
+So if you are running your own git server, you could take advantage of that I guess.
+
+#### TL;DR
+If you are _not_ taking advantage of `prefix` I would recommend `protocolVersion: 1`.
+Otherwise, I recommend to use the default which is `protocolVersion: 2`.
+
 Example Code:
 
 ```js live
+// List all the branches on a repo
 let refs = await git.listServerRefs({
   http,
   corsProxy: "https://cors.isomorphic-git.org",
-  url: "https://github.com/isomorphic-git/isomorphic-git.git"
+  url: "https://github.com/isomorphic-git/isomorphic-git.git",
+  prefix: "refs/heads/",
+});
+console.log(refs);
+```
+
+```js live
+// Get the default branch on a repo
+let refs = await git.listServerRefs({
+  http,
+  corsProxy: "https://cors.isomorphic-git.org",
+  url: "https://github.com/isomorphic-git/isomorphic-git.git",
+  prefix: "HEAD",
+  symrefs: true,
+});
+console.log(refs);
+```
+
+```js live
+// List all the tags on a repo
+let refs = await git.listServerRefs({
+  http,
+  corsProxy: "https://cors.isomorphic-git.org",
+  url: "https://github.com/isomorphic-git/isomorphic-git.git",
+  prefix: "refs/tags/",
+  peelTags: true,
+});
+console.log(refs);
+```
+
+```js live
+// List all the pull requests on a repo
+let refs = await git.listServerRefs({
+  http,
+  corsProxy: "https://cors.isomorphic-git.org",
+  url: "https://github.com/isomorphic-git/isomorphic-git.git",
+  prefix: "refs/pull/",
 });
 console.log(refs);
 ```
