+++ website/versioned_sidebars/version-1.x-sidebars.json
@@ -62,7 +62,8 @@
       "version-1.x-getRemoteInfo2",
       "version-1.x-addRemote",
       "version-1.x-deleteRemote",
-      "version-1.x-listRemotes"
+      "version-1.x-listRemotes",
+      "version-1.x-listServerRefs"
     ],
     "Objects": [
       "version-1.x-readBlob",
